+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Danger - Improve conventional commit scope handling, and non-conventional PR title support ([#105](https://github.com/getsentry/github-workflows/pull/105))
 - Add Proguard artifact endpoint for Android builds in sentry-server ([#100](https://github.com/getsentry/github-workflows/pull/100))
 
 ### Security